backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
anthropic==0.42.0
pydantic==2.10.4
aiosqlite==0.20.0
asyncpg==0.30.0
python-dotenv==1.0.1
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
